+++ example/lib/example5_themes_playground/widgets/panels/theme_simulator/app_example_material3/elevation_screen.dart
@@ -151,7 +151,7 @@ class _ElevationCardState extends State<ElevationCard> {
                 style: Theme.of(context).textTheme.labelMedium,
               ),
               Text(
-                '${widget.info.level} dp',
+                '${widget.info.elevation.toStringAsFixed(0)} dp',
                 style: Theme.of(context).textTheme.labelMedium,
               ),
               if (widget.surfaceTint != null)